DMP-99 action review comments
