+++ config/checkstyle/checkstyle.xml
@@ -27,7 +27,11 @@
     <property name="eachLine" value="true"/>
   </module>
   <module name="BeforeExecutionExclusionFileFilter">
-    <property name="fileNamePattern"  value=".*[\\/]generated[\\/].*$||.*[\\/]controllers[\\/].*$"/>
+    <property name="fileNamePattern"  value=".*[\\/]generated[\\/].*$"/>
+  </module>
+  <module name="SuppressionSingleFilter">
+    <property name="files" value=".*[\\/]controller[\\/].*$" />
+    <property name="checks" value="LineLength" />
   </module>
   <module name="SuppressWarningsFilter" />
   <module name="LineLength">
